Minor changes
'Show Hidden' Menu item is now 'Hide Information' and its usage is
reversed. When selected it will hide information, and when deselected,
it will show information.
Also, when a table entry is being edited while 'Hide Information' is
selected, it will not apply the edit and revert to its previous state.
Changed version number to 3.2.1.